ESP32-C3 RC & Receiver

+++ ESP-IDF_Robot_RC/main/espnow_example_main.c
@@ -99,7 +99,7 @@ void sendData (void) {
     buffer.motor4_rpm_pcm = 0;
     ESP_LOGI(TAG, "x-axis: 0x%04X", (uint8_t)buffer.x_axis);
     ESP_LOGI(TAG, "y-axis: 0x%04X", (uint8_t)buffer.y_axis);
-    ESP_LOGI(TAG, "pcm 1: 0x%04X", buffer.motor1_rpm_pcm);
+    ESP_LOGI(TAG, "pcm 1: 0x%04X", (uint8_t)buffer.motor1_rpm_pcm);
     ESP_LOGI(TAG, "pcm 2: 0x%04X", (uint8_t)buffer.motor2_rpm_pcm);
     ESP_LOGI(TAG, "pcm 3: 0x%04X", (uint8_t)buffer.motor3_rpm_pcm);
     ESP_LOGI(TAG, "pcm 4: 0x%04X", (uint8_t)buffer.motor4_rpm_pcm);
